requirements.txt
async-timeout==4.0.2
bidict==0.22.0
black==23.1.0
cachelib==0.9.0
click==8.1.3
Flask==2.2.1
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.5.1
rel==0.4.8
six==1.11.0
tomli==2.0.1
typing-extensions==4.5.0
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.8.1
zope.event==4.6
zope.interface==5.5.2
